Flushed out a few things
Added common header and footer to all pages. sessions work for login and welcome page. added links to the footer and tested for functionality. not allowed to access the welcome page until you are signed in. added user and pass to the login page for easier testing. added CSS to the header file to control the header and footer easier. Added logo for website.