Major improvements
added functions library, database connections to multiple pages, and proper login screen with functioning SQL backend. Proper create new user page with a functioning SQL backend. Also added the ability to allow admins access to a list of all users in the database. The project still needs products and a cart.  .htaccess has not implemented this commit but there are future plans to add this.